+++ CHANGELOG.md
@@ -11,6 +11,11 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
+- Raise the default mixup and mosaic probability from `0.5` to `0.75` for
+  `ltdetrv2-s/m/l/x` object detection models. Other LTDETR models are unaffected.
+- Lower the default `backbone_lr_factor` from `0.05` to `0.0025` for `ltdetrv2-m/l/x`
+  object detection models. `ltdetrv2-s` and other LTDETR models are unaffected.
+
 ### Deprecated
 
 ### Removed